requirements.txt

Flask>=2.0
Flask-WTF>=1.1.1
python-dotenv>=1.0
